requirements.txt
uvicorn
sqlmodel
pydantic
fastapi
sqlalchemy
python-jose[cryptography]
passlib
starlette
python-multipart
bcrypt==4.1.2